+++ core/src/main/java/io/snabble/sdk/Checkout.java
@@ -182,11 +182,11 @@ public class Checkout {
     }
 
     private PaymentMethod getFallbackPaymentMethod() {
-        // TODO remove project id hack when metadata contains explicit fallback method
-        if(project.getEncodedCodesOptions() != null
-                && !project.getId().contains("ikea")
-                && !project.getId().contains("globus")) {
-            return PaymentMethod.ENCODED_CODES;
+        PaymentMethod[] paymentMethods = project.getAvailablePaymentMethods();
+        for (PaymentMethod pm : paymentMethods) {
+            if (pm.isOfflineMethod()) {
+                return pm;
+            }
         }
 
         return null;